Extract answerEndpoint. Cleanup.
